fix: correct football score update logic and real-time incident handling
- Fixed reducer-side score update structure mismatch
- Adjusted football incident creation flow
- Updated transaction handling in football_tx and tx files
- Synced football incident and score broadcasting logic
- Cleaned and aligned models in database layer
- Updated server initialization for football event flow

+++ database/models/models.go
@@ -164,7 +164,7 @@ type FootballIncident struct {
 	TeamID                *int32    `json:"team_id"`
 	Periods               string    `json:"periods"`
 	IncidentType          string    `json:"incident_type"`
-	IncidentTime          int64     `json:"incident_time"`
+	IncidentTime          int       `json:"incident_time"`
 	Description           string    `json:"description"`
 	PenaltyShootoutScored bool      `json:"penalty_shootout_scored"`
 	CreatedAt             int64     `json:"created_at"`
